refactor home page into separate components
