+++ components/home/SponsoredSection.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { Section } from "../Section";
+import { SectionTitle } from "../../pages/SectionTitle";
+import { SectionProps } from "../../pages/SectionProps";
+
+export const SponsoredSection: React.ElementType = ({
+  content,
+  ...props
+}: SectionProps) => {
+  return (
+    <Section
+      style={{
+        height: 300,
+      }}
+    >
+      <SectionTitle>sponsored content</SectionTitle>
+    </Section>
+  );
+};
