+++ pages/ArticleProps.tsx
@@ -1,0 +1,6 @@
+import { Post } from "../helpers/wpapi";
+
+export interface ArticleProps {
+  post: Post;
+  [key: string]: any;
+}
